missiond/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt r.txt
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
flask==3.0.3
    # via
    #   -r r.txt
    #   flask-cors
    #   flask-login
    #   flask-sqlalchemy
flask-cors==5.0.0
    # via -r r.txt
flask-login==0.6.3
    # via -r r.txt
flask-sqlalchemy==3.1.1
    # via -r r.txt
greenlet==3.0.3
    # via sqlalchemy
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
pillow==10.4.0
    # via -r r.txt
sqlalchemy==2.0.32
    # via flask-sqlalchemy
typing-extensions==4.12.2
    # via sqlalchemy
werkzeug==3.0.4
    # via
    #   flask
    #   flask-login
gunicorn
psycopg2-binary
